requirements.txt
pandas
numpy
scikit-learn
xgboost
pyyaml
streamlit
sqlalchemy
mysql-connector-python
openpyxl
